Refactored the code again and wrote unit tests
