+++ battery/SplinderBattery.py
@@ -1,0 +1,11 @@
+from datetime import date
+from battery.battery import Battery
+
+
+class SplinderBattery(Battery):
+    def __init__(self,last_service_date:date,current_date:date):  
+        self.last_service_date = last_service_date
+        self.current_date = current_date
+
+    def need_service(self) -> bool:
+        return self.current_date - self.last_service_date >= 730
