+++ battery/battery.py
@@ -1,0 +1,5 @@
+from abc import ABC
+
+class Battery(ABC):
+    def need_service(self) -> bool:
+        pass
