+++ engine/engine.py
@@ -1,0 +1,6 @@
+from abc import ABC
+
+
+class Engine(ABC):
+    def needs_service(self):
+        pass
