--- engine/model/calliope.py
@@ -1,12 +1,0 @@
-from datetime import datetime
-
-from engine.capulet_engine import CapuletEngine
-
-
-class Calliope(CapuletEngine):
-    def needs_service(self):
-        service_threshold_date = self.last_service_date.replace(year=self.last_service_date.year + 2)
-        if service_threshold_date < datetime.today().date() or self.engine_should_be_serviced():
-            return True
-        else:
-            return False
